+++ src/sizer/FlexResizer.scss
@@ -1,4 +1,4 @@
-@import './others/variables.scss';
+@import '../others/variables.scss';
 
 .flex-resizer {
   flex: 0 0 4px;
